+++ db/pvault.sql
@@ -55,6 +55,7 @@ CREATE TABLE `formulas` (
   `dilutant` varchar(255) COLLATE utf8_bin DEFAULT NULL,
   `quantity` decimal(8,4) DEFAULT NULL,
   `exclude_from_summary` INT NOT NULL DEFAULT '0', 
+  `exclude_from_calculation` INT NOT NULL DEFAULT '0', 
   `notes` varchar(11) COLLATE utf8_bin DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;